Let ActionForm/index handle which form to choose
