+++ front/app/components/admin/ActionForm/types.ts
@@ -1,5 +1,6 @@
 import { Multiloc } from 'typings';
 
+import { IPermissionData } from 'api/permissions/types';
 import { PermittedBy } from 'api/phase_permissions/types';
 
 export type Changes = {
@@ -8,4 +9,11 @@ export type Changes = {
   verification_expiry?: number | null;
   access_denied_explanation_multiloc?: Multiloc;
   everyone_tracking_enabled?: boolean;
+};
+
+export type Props = {
+  phaseId?: string;
+  permissionData: IPermissionData;
+  onChange: (changes: Changes) => Promise<void>;
+  onReset: () => void;
 };
